# openapi.yaml
openapi: 3.1.0
info:
  title: ReqWise API
  description: API for managing customer project requirements.
  version: 0.1.0
paths:
  /users/register:
    post:
      tags:
      - Users
      summary: Register a new user
      description: Registers a new user with a specified username, email, password,
        role.The password will be hashed before storage. Default role is 'customer'.
      operationId: register_user_users_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /users/token:
    post:
      tags:
      - Users
      summary: User Login
      description: Authenticates a user and returns an OAuth2 access token.
      operationId: login_for_access_token_users_token_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_for_access_token_users_token_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_login_for_access_token_users_token_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: ^password$
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_for_access_token_users_token_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
          description: JWT access token.
        token_type:
          type: string
          title: Token Type
          description: Type of the token (e.g., 'bearer').
          default: bearer
      type: object
      required:
      - access_token
      title: Token
      description: "Schema for the OAuth2 token response.\n\nRepresents the structure\
        \ of the access token returned upon successful login.\n\nAttributes:\n   \
        \ access_token (str): The JWT access token.\n    token_type (str): The type\
        \ of the token, typically \"bearer\"."
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
          title: Username
          description: Unique username for the user.
        email:
          type: string
          format: email
          title: Email
          description: Unique email address for the user.
        full_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Full Name
          description: Optional full name of the user.
        password:
          type: string
          minLength: 8
          title: Password
          description: User's password. Must be at least 8 characters long.
        role:
          $ref: '#/components/schemas/UserRole'
          description: Role of the user (customer or owner).
          default: customer
      type: object
      required:
      - username
      - email
      - password
      title: UserCreate
      description: "Schema for user creation request body.\n\nExtends `UserBase` by\
        \ adding the password and allowing role specification\nduring registration.\n\
        \nAttributes:\n    password (str): User's password. Must be at least 8 characters\
        \ long.\n    role (UserRole): Role of the user (customer or owner). Defaults\
        \ to\n        `UserRole.CUSTOMER`."
    UserOut:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
          title: Username
          description: Unique username for the user.
        email:
          type: string
          format: email
          title: Email
          description: Unique email address for the user.
        full_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Full Name
          description: Optional full name of the user.
        id:
          type: integer
          title: Id
          description: Unique identifier of the user.
        role:
          $ref: '#/components/schemas/UserRole'
          description: Role of the user.
      type: object
      required:
      - username
      - email
      - id
      - role
      title: UserOut
      description: "Schema for user output response body.\n\nExtends `UserBase` by\
        \ including the unique ID and the assigned role,\nsuitable for sending user\
        \ data to clients without sensitive information.\n\nAttributes:\n    id (int):\
        \ Unique identifier of the user.\n    role (UserRole): Role of the user."
    UserRole:
      type: string
      enum:
      - customer
      - owner
      title: UserRole
      description: "Define the possible roles for a user in the system.\n\nAttributes:\n\
        \    CUSTOMER (str): Represents a customer user.\n    OWNER (str): Represents\
        \ a project owner user."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: users/token